View.__add__ to merge_view
verified the cases that used real_strides are redundant
